Change desktop_shop to package format
This changes the entry point from desktop_shop/main.py to
desktop_shop/__main__.py.

As a result of adding an __init__.py package at the root package level,
all imports in the repository changed (which is why this commit contains
so many changes).

The main purpose behind this change was to make the package importable
from outside folders, which is important, e.g. for pypi distribution or
even unit testing with pytest without sys.path hacks.